Refactor PartTable into seperate component
